requirements.txt
greenlet==3.1.1
mysql-connector-python==9.2.0
python-dotenv==1.0.1
SQLAlchemy==1.4.0
